Improve main image greatly.
Improves the main bigrat image greatly by making it a lossless JPEG crop instead of a lossy PNG crop, Increasing the area of the crop to be slightly bigger so more of the tail is seen and also adds a lossless JXL fallforward for newer web browsers. also changes the "perfect" image to match the new crop.