Remove redundant Twitter tags
Twitter will reuse the matching Open Graph tags, and actually doesn't have a URL tag of its own, so it's best to omit them and save bandwidth.

The card property is useful however, and is used by services other than Twitter, so it can be left there. To avoid confusion, I renamed the Open Graph header to Embed Tags.